finsh2fevral

+++ api/models/sale.go
@@ -8,29 +8,27 @@ type Sale struct {
 	Shopassistant_id string    `json:"shopassistant_id"`
 	Cashier_id       string    `json:"cashier_id"`
 	Payment_type     string    `json:"payment_type"`
-	Price            int       `json:"price"`
+	Price            float64   `json:"price"`
 	Status_type      string    `json:"status_type"`
 	Clientname       string    `json:"clientname"`
 	Create_at        time.Time `json:"ctreate_at"`
 }
 type CreateSale struct {
-	Branch_id        string    `json:"branch_id"`
-	Shopassistant_id string    `json:"shopassistant_id"`
-	Cashier_id       string    `json:"cashier_id"`
-	Payment_type     string    `json:"payment_type"`
-	Price            int       `json:"price"`
-	Status_type      string    `json:"status_type"`
-	Clientname       string    `json:"clientname"`
-
+	Branch_id        string  `json:"branch_id"`
+	Shopassistant_id string  `json:"shopassistant_id"`
+	Cashier_id       string  `json:"cashier_id"`
+	Payment_type     string  `json:"payment_type"`
+	Price            float64 `json:"price"`
+	Status_type      string  `json:"status_type"`
+	Clientname       string  `json:"clientname"`
 }
 type UpdateSale struct {
-	ID               string    `json:"id"`
-	Branch_id        string    `json:"branch_id"`
-	Shopassistant_id string    `json:"shopassistant_id"`
-	Cashier_id       string    `json:"cashier_id"`
-	Price            int       `json:"price"`
-	Clientname       string    `json:"clientname"`
-
+	ID               string  `json:"id"`
+	Branch_id        string  `json:"branch_id"`
+	Shopassistant_id string  `json:"shopassistant_id"`
+	Cashier_id       string  `json:"cashier_id"`
+	Price            float64 `json:"price"`
+	Clientname       string  `json:"clientname"`
 }
 
 type SaleRepos struct {
